refactor: Fix routes and structure
Home structured & navbar links fixed
